first try on seperating up the extref compiler passes into more reusable pieces..
